+++ docs/src/componentDocs/AppBar/examples/index.tsx
@@ -1,4 +1,5 @@
 export * from './AppBar';
+export * from './AppBarOverlay';
 export * from './AppBarWithAdditionalContent';
 export * from './CollapsedAppBar';
 export * from './ExpandedAppBar';